refactor: isolate sd-jwt verification tests
